move src/luajit/compat.h → src/game/luajit_compat.h
