+++ src/game/luajit_compat.h
